requirements.txt
PyJWT
pygithub
python-dotenv
github3.py
Django==2.2.4
selenium
pyyaml
social-auth-app-django
timeago
urllib3